+++ src/Contracts/RouteStatisticInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Bilfeldt\LaravelRouteStatistics\Contracts;
-
-interface RouteStatisticInterface
-{
-    public static function incrementOrCreate(array $attributes, int $amount = 1);
-}