+++ tests/Unit/RequestMacroTest.php
@@ -11,23 +11,12 @@ class RequestMacroTest extends TestCase
     public function test_enables_route_statistics()
     {
         $request = new Request();
-        $stats = LaravelRouteStatisticsFacade::disable();
 
-        $this->assertFalse($stats->isEnabled());
+        $this->assertEmpty($request->attributes->get('log'));
 
         $request->routeStatistics();
 
-        $this->assertTrue($stats->isEnabled());
-    }
-
-    public function test_merges_attributes()
-    {
-        $request = new Request();
-        $stats = LaravelRouteStatisticsFacade::disable();
-        $stats->setAttributes(['first_name' => 'john']);
-
-        $request->routeStatistics(['last_name' => 'doe']);
-
-        $this->assertEquals(['first_name' => 'john', 'last_name' => 'doe'], $stats->getAttributes());
+        $this->assertIsArray($request->attributes->get('log'));
+        $this->assertContains('routestat', $request->attributes->get('log'));
     }
 }